Reapply "Split AiPlayerTest into speak/choose/cross-cutting test files"
This reverts commit 750d8eb8f491243ad8dad924a8d71763728d2a71.
